fixed scaling issue with iPhone 6 plus

+++ platforms/ios/Gearboy/Emulator.h
@@ -41,7 +41,7 @@
     BOOL firstFrame;
 }
 
-@property (nonatomic) int multiplier;
+@property (nonatomic) float multiplier;
 @property (nonatomic) BOOL retina;
 @property (nonatomic) BOOL iPad;
 @property (nonatomic) GLKView* glview;
@@ -62,7 +62,7 @@
 - (void)renderFrame;
 - (void)renderMixFrames;
 - (void)setupTextureWithData: (GLvoid*) data;
-- (void)renderQuadWithViewportWidth: (int)viewportWidth andHeight: (int)viewportHeight andMirrorY: (BOOL)mirrorY;
+- (void)renderQuadWithViewportWidth: (float)viewportWidth andHeight: (float)viewportHeight andMirrorY: (BOOL)mirrorY;
 - (void)setAudio: (BOOL)enabled;
 - (void)resetAudio;
